# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /groups:
    get:
      operationId: listGroups
      tags: [groups]
      summary: List the groups I belong to
      responses:
        "200":
          description: The caller's groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Group"
    post:
      operationId: createGroup
      tags: [groups]
      summary: Create a new group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
  /groups/{id}:
    get:
      operationId: getGroup
      tags: [groups]
      summary: Get a group's details and members
      description: Only members of the group may read it.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Group details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupDetail"
        "403":
          description: Not a member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: renameGroup
      tags: [groups]
      summary: Rename a group (owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
        "403":
          description: Not the owner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /groups/{id}/leave:
    post:
      operationId: leaveGroup
      tags: [groups]
      summary: Leave a group
      description: >-
        Removes the caller from the group. If the owner leaves, ownership passes
        to the longest-standing remaining member; if no members remain, the
        group is deleted.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Left
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /groups/{id}/invite:
    post:
      operationId: createOrGetGroupInvite
      tags: [invites]
      summary: Create or fetch the group's shareable invite link
      description: >-
        Returns the group's reusable invite link, creating one on first call.
        Only members may generate it.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: The group's invite
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invite"
        "403":
          description: Not a member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /invites/{token}:
    get:
      operationId: getInvitePreview
      tags: [invites]
      summary: Look up an invite by token (public)
      description: >-
        Public endpoint used by the invite landing page to show which group the
        visitor is being invited to before they sign up. Returns valid=false for
        unknown tokens.
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Invite preview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvitePreview"
  /invites/{token}/accept:
    post:
      operationId: acceptInvite
      tags: [invites]
      summary: Accept an invite after signing in
      description: >-
        Adds the authenticated caller to the invite's group. Idempotent;
        accepting when already a member is a no-op.
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Acceptance result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AcceptInviteResult"
        "404":
          description: Invite not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Invite:
      type: object
      required:
        - token
        - groupId
        - groupName
        - createdAt
      properties:
        token:
          type: string
        groupId:
          type: integer
        groupName:
          type: string
        createdAt:
          type: string
          format: date-time
    InvitePreview:
      type: object
      required:
        - token
        - groupName
        - inviterName
        - valid
      properties:
        token:
          type: string
        groupName:
          type: string
        inviterName:
          type: string
        valid:
          type: boolean
    AcceptInviteResult:
      type: object
      required:
        - joined
        - groupId
        - groupName
      properties:
        joined:
          type: boolean
        groupId:
          type: ["integer", "null"]
        groupName:
          type: string
    GroupInput:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 60
    Group:
      type: object
      required:
        - id
        - name
        - ownerId
        - role
        - memberCount
        - createdAt
      properties:
        id:
          type: integer
        name:
          type: string
        ownerId:
          type: string
        role:
          type: string
          enum: [owner, member]
        memberCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    GroupMember:
      type: object
      required:
        - userId
        - displayName
        - role
        - joinedAt
      properties:
        userId:
          type: string
        displayName:
          type: string
        role:
          type: string
          enum: [owner, member]
        joinedAt:
          type: string
          format: date-time
    GroupDetail:
      type: object
      required:
        - id
        - name
        - ownerId
        - role
        - memberCount
        - createdAt
        - members
      properties:
        id:
          type: integer
        name:
          type: string
        ownerId:
          type: string
        role:
          type: string
          enum: [owner, member]
        memberCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        members:
          type: array
          items:
            $ref: "#/components/schemas/GroupMember"
...