# lib/api-zod/src/generated/types/groupDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GroupDetailRole } from './groupDetailRole';
import type { GroupMember } from './groupMember';

export interface GroupDetail {
  id: number;
  name: string;
  ownerId: string;
  role: GroupDetailRole;
  memberCount: number;
  createdAt: Date;
  members: GroupMember[];
}