# lib/api-zod/src/generated/types/groupMember.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GroupMemberRole } from './groupMemberRole';

export interface GroupMember {
  userId: string;
  displayName: string;
  role: GroupMemberRole;
  joinedAt: Date;
}